requirements.txt
cachelib==0.2.0
certifi==2020.12.5
chardet==4.0.0
click==8.0.1
Flask==2.0.1
Flask-Cors==3.0.10
Flask-Session==0.4.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.0
idna==2.10
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
memcache==0.8.0
PyYAML==3.13
redis==3.5.3
requests==2.25.1
six==1.16.0
SQLAlchemy==1.4.22
urllib3==1.26.3
Werkzeug==2.0.1
